Backend AI part/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-google-genai
langchain-core
supabase
python-dotenv
google-generativeai